make clear which bar belongs to which answer (#16)
it is unclear where a bar belongs to if the space to next/previous answer is the same. does the bar belong to white or to blue? you need to look at first/last row to figure that out:

![Screenshot 2025-04-11 at 13.58.45.png](/attachments/37650cf7-293e-4800-8788-c1f202888797)

better:

![Screenshot 2025-04-11 at 14.05.17.png](/attachments/1fe6dc1c-082e-4369-8533-aab519568ca3)

it was like that before multiple answers were introduced, btw

moreover, the PR updates the README, and makes labels tappable

Reviewed-on: https://codeberg.org/webxdc/poll/pulls/16
Co-authored-by: B. Petersen <r10s@b44t.com>
Co-committed-by: B. Petersen <r10s@b44t.com>